+++ qbraid/runtime/result.py
@@ -26,7 +26,7 @@ from qbraid_core.system.generic import _datetime_to_str
 
 from .enums import ExperimentType
 from .postprocess import counts_to_probabilities, normalize_counts
-from .schemas.experiment import GateModelExperimentMetadata
+from .schemas.experiment import AnnealingExperimentMetadata, GateModelExperimentMetadata
 
 KeyType = TypeVar("KeyType", str, int)
 
@@ -266,6 +266,54 @@ class AhsResultData(ResultData):
             "measurement_counts": self._measurement_counts,
             "measurements": self._measurements,
         }
+
+
+class AnnealingResultData(ResultData):
+    """Class for storing and accessing the results of an annealing job.
+
+    Args:
+        ResultData (ResultData): ABC for runtime results linked to a specific ExperimentType.
+
+    Returns:
+        AnnealingResultData: An instance of AnnealingResultData.
+    """
+
+    def __init__(self, sa_results: Optional[dict[str, Any]] = None):
+        """Create a new AnnealingResultData instance.
+
+        Args:
+            sa_results (Optional[dict[str, Any]], optional):
+                Ex: [{'spin': {' x1': 0, ' x2': 0, 'x1': 0},\
+                'energy': 0, 'time': 0.006517000030726194,\
+                'constraint': True, 'memory_usage': 1.189453125}]. 
+                Defaults to None.
+        """
+        self._sa_results = sa_results or []
+
+    @property
+    def experiment_type(self) -> ExperimentType:
+        """Returns the experiment type."""
+        return ExperimentType.ANNEALING
+
+    @classmethod
+    def from_object(cls, model: AnnealingExperimentMetadata, **kwargs) -> AnnealingResultData:
+        """Creates a new AnnealingResultData instance from a AnnealingExperimentMetadata object."""
+        return cls.from_dict(model.model_dump(**kwargs))
+
+    @classmethod
+    def from_dict(cls, data: dict[str, Any] = None) -> AnnealingResultData:
+        """Creates a new AnnealingResultData instance from a dictionary."""
+        return cls(sa_results=data.get("sa_results"))
+
+    def to_dict(self) -> dict[str, Any]:
+        """Converts the AnnealingResultData instance to a dictionary."""
+        return {
+            "sa_results": self._sa_results,
+        }
+
+    def get_results(self) -> dict[str, Any]:
+        """Returns the job data."""
+        return self._sa_results
 
 
 class Result: